refactor code add todo a not dumb thing for the "to"
