+++ src/test/scala/fr/charlotte/RuntimeTest.scala
@@ -2,19 +2,20 @@ package fr.charlotte
 import org.scalatest.funsuite.AnyFunSuite
 import fr.charlotte.ast.Parser
 import fr.charlotte.lexing.Lexer
-import fr.charlotte.ast.Translator
-class TranslatorTest extends AnyFunSuite {
+import fr.charlotte.runtime.Translator
+
+class RuntimeTest extends AnyFunSuite {
   test("Test Or Database for STAR"){
-    val p = Translator("lignes",Parser(Lexer("C1 or C2").lex()).parse().tpe).translate
+    val p = runtime.Translator("lignes",Parser(Lexer("C1 or C2").lex()).parse().tpe).translate
     val should = "((lignes like \"%C1%\") OR (lignes LIKE \"%C2%\"))"
     assert(p.equalsIgnoreCase(should))
   }
   test("Nested And Database for STAR"){
-    val p = Translator("lignes", Parser(Lexer("C1 and C2 and C3").lex()).parse().tpe).translate
+    val p = runtime.Translator("lignes", Parser(Lexer("C1 and C2 and C3").lex()).parse().tpe).translate
     val should = "(((lignes LIKE \"%C1%\") AND (lignes LIKE \"%C2%\")) AND (lignes LIKE \"%C3%\"))"
     assert(p.equalsIgnoreCase(should))
   }
-  
-  
-  
+
+
+
 }
