+++ src/test/scala/fr/charlotte/lexing/LexerTest.scala
@@ -1,8 +1,7 @@
-package fr.charlotte
+package fr.charlotte.lexing
 
+import fr.charlotte.lexing.{Lexer, Token}
 import org.scalatest.funsuite.AnyFunSuite
-import fr.charlotte.lexing.Lexer
-import fr.charlotte.lexing.Token
 class LexerTest extends AnyFunSuite:
   test("OrTest"){
     val result = Lexer("(C1 or C2)").lex()
