+++ src/test/scala/fr/charlotte/parsing/ParserTests.scala
@@ -1,14 +1,10 @@
-package fr.charlotte
+package fr.charlotte.parsing
 
-
-import org.scalatest.funsuite.AnyFunSuite
-import fr.charlotte.lexing.Lexer
-import fr.charlotte.lexing.Token
-import fr.charlotte.ast.Parser
-import fr.charlotte.ast.Ast
 import fr.charlotte.ast.Ast.Tree.*
-import fr.charlotte.ast.Parameter
 import fr.charlotte.ast.Parameter.Type.*
+import fr.charlotte.ast.{Ast, Parameter, Parser}
+import fr.charlotte.lexing.{Lexer, Token}
+import org.scalatest.funsuite.AnyFunSuite
 
 class ParserTests extends AnyFunSuite{
   test("Simple command test"){
